requirements.txt
fastapi==0.109.2
pydantic==2.5.3


uvicorn==0.27.0
sqlalchemy==2.0.25
jinja2==3.1.3
openai==1.10.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
alembic==1.13.1
psycopg2